Setting up Handling Exception

+++ Application/Activities/Delete.cs
@@ -1,3 +1,4 @@
+using Application.Core;
 using MediatR;
 using Persistence.Data;
 
@@ -5,12 +6,12 @@ namespace Application.Activities
 {
     public class Delete
     {
-        public class Command : IRequest
+        public class Command : IRequest<Result<Unit>>
         {
             public Guid Id { get; set; }
         }
 
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
         private readonly DataContext _context;
             public Handler(DataContext context)
@@ -18,14 +19,17 @@ namespace Application.Activities
             _context = context;
             }
 
-            public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var activity = await _context.Activities.FindAsync(request.Id);
 
-                _context.Activities.Remove(activity);
-                await _context.SaveChangesAsync();
+                if(activity == null) return null;
 
-                return Unit.Value;
+                _context.Activities.Remove(activity);
+                var result = await _context.SaveChangesAsync() > 0;
+                if(!result) return Result<Unit>.Failure("Failed to delete activity");
+
+                return Result<Unit>.Success(Unit.Value);
             }
         }
     }
